cleanup and relocation per @rakuy0
